LayoutTests/media/media-source/media-source-first-sample-pts-non-zero-canplay-expected.txt

This test verifies that the "canplay" event is dispatched on a video element with a MediaSource that loads a video track whose first sample has a presentation timestamp greater than 0.
EVENT(sourceopen)
END OF TEST
